move the traits to the class as it is only used once
